Separate AMP preview button into its own component
