+++ assets/css/src/amp-block-editor.css
@@ -4,7 +4,7 @@
 }
 
 /* AMP preview button wrapper */
-.wp-core-ui #amp-wrapper-post-preview {
+.wp-core-ui .amp-wrapper-post-preview {
 	margin-left: -6px;
 	margin-right: 6px;
 }
